PL5/Sesion3_2_y_Sesion3_3/pyproject.toml
[project]
name = "sesion3-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.135.3",
    "fastapi-cli>=0.0.24",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "sqlmodel>=0.0.38",
]
